refactor: move 13 views to shared-web, extract planner types to entity, fix BCE violation in ProofOfWorkView
Phase 6.9 completion: resolve 4 blockers preventing view migration to shared-web module.

- Blocker A: Extract PlannerPlanPreview + PlannerPreviewState from orchestration.control
  to orchestration.entity (BCE fix — pure data was in control layer). Add planDomain dep
  to orchestrationDomain.
- Blocker C: Refactor ProofOfWorkView.panel to accept pre-computed RequirementCheck list
  instead of calling ProofOfWorkExtractor directly (BCE fix — boundary was calling control
  logic). Push validation up to IssueController.
- Blocker D: Add llm4zio + orchestrationDomain to sharedWeb dependencies.
- Move 13 view files (6,916 lines) from root shared/web to shared-web module.

Root: 158 → 145 source files. shared-web: 34 → 46 view files.
Only CheckpointsView remains in root (blocked by CheckpointSnapshot → GatewayConfig chain).
