+++ modules/shared-web/src/main/scala/shared/web/AgentMonitorView.scala
