+++ modules/shared-web/src/main/scala/shared/web/ChatView.scala
