+++ modules/shared-web/src/main/scala/shared/web/CommandCenterView.scala
