+++ modules/shared-web/src/main/scala/shared/web/ProjectsView.scala
