+++ modules/shared-web/src/main/scala/shared/web/SettingsView.scala
